Add a happy-path test for get_diff_from_repo.
As part of this, I added a tar of a git repo for testing purposes. The tar isn't compressed, because it's very small, should change infrequently, and I care more about test performance than repo space for this tiny object.

I also added a file that represents the expected output, which I actually generated by running "git diff" and piping it to a file, so that is exactly what git diff produces in that situation.